capsules/extra/nonvolatile_storage_driver: remove unnecessary usize cast

+++ capsules/extra/src/nonvolatile_storage_driver.rs
@@ -457,7 +457,7 @@ impl hil::nonvolatile_storage::NonvolatileStorageClient for NonvolatileStorage<'
                                 read.mut_enter(|app_buffer| {
                                     let read_len = cmp::min(app_buffer.len(), length);
 
-                                    let d = &app_buffer[0..(read_len as usize)];
+                                    let d = &app_buffer[0..read_len];
                                     for (i, c) in buffer[0..read_len].iter().enumerate() {
                                         d[i].set(*c);
                                     }
